# backend/docs/swagger.yaml
basePath: /api/
definitions:
  api.HostMetadataPatchRequest:
    properties:
      location:
        type: string
      notes:
        type: string
      owner:
        type: string
      pinned:
        type: boolean
      tags:
        items:
          type: string
        type: array
    type: object
...
  models.Host:
    properties:
      date:
        type: string
      deviceType:
        type: string
      dns:
        type: string
      hw:
        type: string
      id:
        type: integer
      iface:
        type: string
      ip:
        type: string
      known:
        type: integer
      location:
        type: string
      mac:
        type: string
      name:
        type: string
      notes:
        type: string
      now:
        type: integer
      owner:
        type: string
      pinned:
        type: boolean
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    url: https://github.com/godlev/LANnventory
  description: Local network inventory and monitoring application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LANnventory API
  version: 0.1.0-beta.2
paths:
...
  /export/backup:
    get:
      description: Export a versioned logical backup containing current hosts, host
        history, activity events and host metadata. Configuration secrets are not
        included.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Backup generation failure
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download portable data backup
      tags:
      - export
  /export/inventory.csv:
    get:
      description: Export the current device inventory and metadata as CSV. This is
        not a full backup and does not include history or events.
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: CSV generation failure
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download current inventory CSV
      tags:
      - export
...
  /history/{mac}:
    get:
      description: |-
        Retrieve the latest history entries for a specific host by MAC address
        Inventory metadata fields are not included on history rows.
      parameters:
      - description: MAC address of the host
        in: path
        name: mac
        required: true
        type: string
      - description: Number of history entries to return
        in: query
        name: num
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Host'
            type: array
      summary: Get history by MAC
      tags:
      - history
  /history/{mac}/{date}:
    get:
      description: |-
        Retrieve history for a specific host on a given date
        Inventory metadata fields are not included on history rows.
        The date format is flexible and can be:
        - Year only: `2025`
        - Year + month: `2025-09`
        - Full date: `2025-09-06`
        - Full timestamp: `2025-09-06 00:58:26`
      parameters:
      - description: MAC address of the host
        in: path
        name: mac
        required: true
        type: string
      - description: Date filter (supports YYYY, YYYY-MM, YYYY-MM-DD, YYYY-MM-DD HH:mm:ss)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Host'
            type: array
      summary: Get history by date
      tags:
      - history
...
  /host/{id}/metadata:
    patch:
      consumes:
      - application/json
      description: Partially update manually managed inventory metadata. Tags are
        trimmed, empty tags are removed, duplicates are removed case-insensitively,
        and user order is preserved.
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      - description: Metadata payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.HostMetadataPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update host metadata
      tags:
      - hosts
...
swagger: "2.0"