# backend/internal/models/models.go
package models

// Conf - app config
type Conf struct {
	Host                  string
	Port                  string
	Theme                 string
	Color                 string
	DirPath               string
	ConfPath              string
	DBPath                string
	NodePath              string
	LogLevel              string
	Ifaces                string
	ArpArgs               string
	ArpStrs               []string
	Timeout               int
	TrimHist              int
	ConnectivityRetention int
	ShoutURL              string
	Version               string
	// PostgreSQL
	UseDB     string
	PGConnect string
	// InfluxDB
	InfluxEnable  bool
	InfluxAddr    string
	InfluxToken   string
	InfluxOrg     string
	InfluxBucket  string
	InfluxSkipTLS bool
	// Prometheus
	PrometheusEnable bool
}

// Host - one host
type Host struct {
	ID         int      `gorm:"column:ID;primaryKey"`
	Name       string   `gorm:"column:NAME"`
	DNS        string   `gorm:"column:DNS"`
	Iface      string   `gorm:"column:IFACE"`
	IP         string   `gorm:"column:IP"`
	Mac        string   `gorm:"column:MAC"`
	Hw         string   `gorm:"column:HW"`
	Date       string   `gorm:"column:DATE"`
	Known      int      `gorm:"column:KNOWN"`
	Now        int      `gorm:"column:NOW"`
	DeviceType string   `gorm:"column:DEVICE_TYPE"`
	Owner      string   `gorm:"-"`
	Location   string   `gorm:"-"`
	Notes      string   `gorm:"-"`
	Tags       []string `gorm:"-"`
	Pinned     bool     `gorm:"-"`
	// MetadataLoaded marks current-host API responses that were explicitly enriched.
	MetadataLoaded bool `gorm:"-" json:"-"`
}

// Stat - status
type Stat struct {
	Total   int
	Online  int
	Offline int
	Known   int
	Unknown int
}